+++ snappylapy/models.py
@@ -3,8 +3,38 @@
 from __future__ import annotations
 
 import pathlib
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from snappylapy.constants import DIRECTORY_NAMES
+
+
+@dataclass
+class DependingSettings:
+    """Settings for depending on other snapshots. Used for loading snapshots."""
+
+    test_filename: str
+    """Filename of the test module where the depending test are defined."""
+
+    test_function: str
+    """Name of the depending test function."""
+
+    snapshots_base_dir: pathlib.Path
+    """Input base directory for loading snapshots."""
+
+    filename_extension: str | None = None
+    """Extension of the depending snapshot file."""
+
+    custom_name: str | None = None
+    """Custom name for the depending snapshot file."""
+
+    @property
+    def filename(self) -> str:
+        """Get the depending snapshot filename."""
+        if not self.filename_extension:
+            msg = "Missing depending snapshot filename extension."
+            raise ValueError(msg)
+        if self.custom_name is not None:
+            return f"[{self.test_filename}][{self.test_function}][{self.custom_name}].{self.filename_extension}"
+        return f"[{self.test_filename}][{self.test_function}].{self.filename_extension}"
 
 
 @dataclass
@@ -30,17 +60,12 @@ class Settings:
     """Extension for the output of snapshot file."""
 
     # Configurations for depending
-    depending_test_filename: str | None = None
-    """Filename of the test module where the depending test are defined. Used for loading."""
+    depending_tests: list[DependingSettings] = field(default_factory=list)
+    """
+    Depending tests are used for loading snapshots from other tests.
 
-    depending_test_function: str | None = None
-    """Name of the depending test function. Used for loading."""
-
-    depending_filename_extension: str | None = None
-    """Extension of the depending snapshot file. Used for loading."""
-
-    depending_snapshots_base_dir: pathlib.Path | None = None
-    """Input base directory for loading snapshots."""
+    Information about each test the users have specified in a test decorator will be stored here.
+    """
 
     @property
     def snapshot_dir(self) -> pathlib.Path:
@@ -58,17 +83,3 @@ class Settings:
         if self.custom_name is not None:
             return f"[{self.test_filename}][{self.test_function}][{self.custom_name}].{self.filename_extension}"
         return f"[{self.test_filename}][{self.test_function}].{self.filename_extension}"
-
-    @property
-    def depending_filename(self) -> str:
-        """Get the depending snapshot filename."""
-        if (
-            not self.depending_test_filename
-            or not self.depending_test_function
-            or not self.depending_filename_extension
-        ):
-            msg = "Missing depending test filename, function or extension."
-            raise ValueError(msg)
-        if self.custom_name is not None:
-            return f"[{self.depending_test_filename}][{self.depending_test_function}][{self.custom_name}].{self.depending_filename_extension}"  # noqa: E501
-        return f"[{self.depending_test_filename}][{self.depending_test_function}].{self.depending_filename_extension}"